Add backend API to get content for routing
