+++ pages/api/content/getContentById.ts
@@ -8,12 +8,10 @@ export default async function handler(
 ) {
     // connect to the database
     await ConnectMongo();
-    // get the content id
-    const contentId = req.body.id;
-    const content = await Content.findOne({
-        _id: contentId,
-    });
-    console.log(content);
+    // get the content id from the query param
+    const contentId = req.query.id;
+    console.log(contentId);
+    const content = await Content.findById(contentId);
     if (!content) {
         res.status(401).json({ name: "Could not find content" });
         return;
